Mac Carbon 3.8.10b4 MAXPATHLEN is defined in the os-x headers as 1024, versus in the unix vm as 1000, so switch to DOCUMENT_NAME_SIZE which is 1000.
Alter sqMacUnixFileInterface not to aliases resolve the last part of path we wanderDown because it might be the alias we want, versus the target of the alias. 

Fix SqMacUIAppleEvents to use posix path name, versus attempting to translate posix path name into a posix path name which causes double click on image but vm already running applescript launch of another VM to fail.