CogVM source as per VMMaker.oscog-eem.114.
CoInterpreter: fix bad bug with primitiveClone of a compiled method.
The clone should be an uncogged method irrespective of the state of
the receiver.  Symptoms include rare crashes in become: during
MethodDictionaryTest which copies a test method.

Cogit: Fix potential bug in become: with cogged methods.  Since
become can cause an object to gain a new reference cogged methods
that gain new references must be added to youngReferrers.