CogVM source as per VMMaker.oscog-eem.117. Fix bugs described in the "[Pharo-project] Troubles with #flushCache and #run:with:in:" thread http://lists.gforge.inria.fr/pipermail/pharo-project/ 2011-July/050858.html.  The PIC machinery wrongly treated invoke-as-method sends as MNUs.
Closely related, finally fully implement PIC MNU cacheing where, by calling a
special abort, a PIC is able to record that a gven selector is an MNU for a
particular class.  Speeds up a simple MNU benchmark by 33% (with more
performance the deeper the receiver's cass hierarchy is).

On Mac fix mis-editing of Info.plist to insert revision info so that VM .app
once again starts on 10.5.x.

Update Pharo icon to the nice Cog one.