CogVM source as per VMMaker.oscog-eem.156
CoInterpreter:
Provide a thorough flush primitive for CompiledMethods that discards all
machine code and makes sure that any contexts using the method have bytecode
pcs.  Primitive #215 (same as 116 in the Stack VM).  This is much slower than
116 (flushCache) since it has to enumerate over all heap contexts.

Provide an xray primitive for CompiledMethod that answers if a method has
machine code, and if so if it's machine code is frameless, and/or refers
to a young object.  No primitive number.  Used to test the above.

Make printOopShort: print Association keys.  Useful for
longPrintOop:, and hence printReferencesTo: etc.

Mac OS: add fflush to debug printing in sqMacUIEventsUniversal.c so output
appears promptly.

Build: a few more svn:ignore properties.