CogVM source as per VMMaker.oscog-eem.255.
[New[Co]]ObjectMemory:
Fix becomeForward: so that objects whose references are deleted are
freed and can no longer be resurrected via allObjects or allInstances.
Remove freed young roots from the rootsTable.  Filter freed objects
pointed to from the extraRootsTable (because these locations can change
it is wrong to remove entries from the extraRootsTable).

Make primitiveIdentityHash pop all arguments, for Newspeak VMMirrors.

StackToRegisterMappingCogit:
Fix marshalling of absent receiver sends.  The items beneath the
arguments (and to-be-pushed receiver) must be spilled before
the receiver is pushed.

Improve code quality for numArgs > numRegArgs sends when
receiver is not a spill and there are no uses of ReceiverResultReg
amongst args.  e.g. now avoids loading ReceiverResultReg from stack
in code such as 1 with: 2 with: 3 .