CogVM source as per VMMaker.oscog-eem.261.
Move determination of the ammount of headroom to the platform in
osCogStackPageHeadroom (in the various sqFooMain.c files).

Provide a routine to monitor the ammount of unused headroom,
which requires the stack memory be zeroed before use.  Assume
the platform will provide a -reportheadroom flag for enabling the
report.  Provide primitiveMinimumUnusedHeadroom for in-image
access.

Add some asserts to check that a page's frame pointer is always
in range (setHeadFP:andSP:inPage: already did this).

Add the GdbARMPlugin.

Correct cygwin HowToBuilds.

Nuke the sparseimage in the Newspeak mac installer.

Add the named serial port read & write primitives to the SerialPlugin.

Fix change history upload in uploadvms script.