CogVM source as per VMMaker.oscog-eem.333
Speed-up Newspeak significantly (e.g. -28% in one compile-intensive benchmark)
by maintaining unpaired methods (compilations of anonymous accessors) on a
linked list instead of searching the entire method zone.

Do this by adding a NewspeakCogMethod and surrogates that add a nextMethod
field.  Update surrogate accessor generation so that accessors can answer and
receive surrogates and nils.  Also add a nop before the dynSuperEntry so as to
change its alignment now that (in the Newspeak VM) CogMethod has changed size
in gaining a field.

Eliminate CogStackPage, collapsing it onto InterpreterStackPage
(it only has a class side).

Use temporaryCountOfMethodHader: instead of tempCountOf: in
SimpleStackBasedCogit>>compileFrameBuild (header is in hand).