CogVM source as per VMMaker.oscog-eem.545.
Fix setting of endOfMemory on load in SqueakV3 VMs.

Spur:

Slight streamlining in followForwardingPointersInStackZone:, plus fixingfollowjng of frameMethodObject vs frameMethod in machine-code frames.

Fix the concatenation of the assert-failure reason in
isValidObjStackPage:myIndex:firstPage: so that it works in C.

Slang:
Generate the CoInterpreter *before* the Cogit so that the objectMemory andinterpreter classes have initialized *before* the Cogit is generated.  Otherwisethe Cogit can be generated with values from the wrong interpreter/objectMemory.