CogVM source as per VMMaker.oscog-eem.624
Comment and coordinate the various examplePlugins.int/ext files.

Spur:
Instance printing should not print free chunks ;-)

Remember to invoke markAndTraceOrFreeMachineCode: in the mark phase of full GC.

Add classOrNilAtIndex: and use it in places where a nil return is unacceptable
(fetchClassOf:, especially for isKindOfClass: when given garbage in printOop:).

Improve spread in Spur's identity hash from 64% to 75%.
Make primitiveSizeInBytesOfInstance accept 2 args in Newspeak for VMMirror.

Use followForward: when accessing the elements of the array(s) in become, since
there's nothing that ensures uniqueness and/or disjointness in the elements.
Hence explicitly filter-out an object becomming itself.