CogVM source as per VMMaker.oscog-eem.638
Add primitiveAllInstances and modify primitiveAllObjects so that
both will grow memory to guarantee their results.
Modify Objectmemory/SpurMemoryManager>>allObjects &
allInstancesOf: to answer the count of objects when they run out of
space so that the primitives can grow and retry.

Change the time primtiives to access the time now, not the time as
updated by the heartbeat (but /don't/ change the time basis for
event checking.  This for performance because use of gettimeofday
in e.g. stack overflow can be a signficant performance overhead.
Move the simulated time implementations (ioSeconds et al) up to
StackInterpreter to reduce duplication.  Add ioSecondsNow et al to
the platform code to support this.

Spur:
Assert that the markStack and weaklingStack are empty at system
startup and at start of GC (allInstancesOf: et al went though a buggy
phase where weaklingStack was left full).

Fix assert in popObjStack: now that isValidObjStack: checks slots in
hiddenRootsObj.