CogVM source as per VMMaker.oscog-eem.673
Sista:
We cannot back-up the machine-code pc in ceCounterTripped: because there isn't
always a send preceeding the jump (e.g. in and: [] and or: [] chains).  Instead,
prevent subsequent counter trips (to reduce the number of necessary
skipBackBeforeJump sends in the image) by nilling out SelectorCounterTripped. The
image-level optimization code will both restore the selector and back-up the pc.