CogVM source as per VMMaker.oscog-eem.674
Deal with the pc ambiguity of conditional branches followed by backward branches
(conditional branches map to the following pc, backward branches to themselves)
by providing and using mcPCForBackwardBranch:startBcpc:in: when converting
interpreted frames containing loops to machine code frames.  This eliminates a
bogus mustBeBoolean when converting interpreter loops in Sista.

Change mcPCFor[BackwardBranch]:startBcpc:in: to answer the absolute mcpc (as was
commented) rather than the relative mcpc.

This requires a 5-argument perform in mapFor:bcpc:performUntil:arg:.  Refactor
to add an isBackwardBranch argument to the function perfomed in all clients.

For performance, specify that mapFor:[bcpc:]performUntil:arg: are
inlined, eliminating the perform/indirect function call.  This adds of
the order of 3% to the size of a cogit.o's text seg so is acceptable.

Fix a bug in printing frame flags (order was the wrong way around).