CogVM source as per VMMaker.oscog-djm.688
ARM32FFIPlugin:
Change dummyFloatFunction to loadFloatRegs since it is a better name. Define this
function elsewhere (platforms/Cross/plugins/SqueakFFIPrims/sqFFIPlugin.c) so that
the C optimizer cannot optimize it away. This fixes a problem when calling out to
functions taking float parameters when compiling with optimization level > 0.

VMMaker image:
Add an expression to the workspace to generate plugins.