CogVM source as per VMMaker.oscog-eem.706
ARM FFI: Fix bug in passing structure argument in the case where a struct would
normally be split between registers and stack, but because of prior coprocessor
register candidiate args having been pushed on the stack, the structure should
not be split but pushed entirely on the stack.

Spur:
Fix the coallescing sweep in pig compact to start from the
lowest of the first forwarder or the first free chunk.

Do a tenuring scavenge to void eden in freeUnmarkedObjectsAndSortAndCoalesce-
FreeSpaceForPigCompact to avoid any tenuring due to futureSpace overflow in the
unmarking scavenge in eliminateAndFreeForwardersForPigCompact.  Do this by
adding the MarkOnTenure tenuring criterion.  Change the use of shouldBeTenured:
to allow it to be inlined into copyAndForward:.

Add an assert to freeChunkWithBytes:at: to catch accidental freeing of
non-oldSpace.  Refactor SpurSegmentManager>>postSnapshot to not fall foul of
this new assert.  Add a check to totalFreeListBytes to catch double freeing
(double freeing was caused by tenuring in the eAFFFPC scavenge growing memory to
make room for tenuring, given free lists are not rebuilt at this point, and
later in eAFFFPC this memory being freed again).

Don't inline doScavenge: or scavengingGCTenuringIf:.

Refactor eliminateAndFreeForwardersForFitCompact to extract the large loops.

Add printForwarders for debugging.

With these changes the StackInterpreter runs the Newspeak bootstrap to
completion, albeit with one presumably bogus assert fail to do with a
DependentsArray referring to forwarder contents.

Cogit:
Fix a double free bug in unlinkSendsOf:isMNUSelector:.  Harmless cuz the result
is only a bogus count of how many methods freed.  Beef up the cog method
integrity check to verify a its methodObject is a CompiledMethod.

Plugins:
Regenerate some that contain formatting changes due to VMMaker.oscog-eem.702/701.