CogVM source as per VMMaker.oscog-eem.733
Use unsigned comparisons when testing if there's sufficient heap space in
readImageFromFile:HeapSize:StartingAt:.  Leads to accurate failure message when
failing to allocate large heap.

Add an assert to check that all weaklinks are accessible (but the bug has
already been found with the nextCorpse encoding bug).

Spur:
Implement primitives to get (primitiveIsPinned) and (un)set
(primitivePin) per-object pinning.

Fix traversal in sortedFreeListPairwiseReverseDo: when there is only
one element in the sorted free list.

Weak containers and ephemerons in new space must be processed after all strongly
reachable objects have been scavenged.  So when the scavenger reaches a weak
container or ephemeron it copies it to future space, leaving behind a forwarding
pointer in the corpse, but it does not yet scavenge its contents.  The corpse is
used to construct the list of scavenged weak/ephemeral objects.  The corpse's
forwarding pointer is already taken to point to the survivor.  Instead, the list
is constructed by having a list head variable, and linking the list through a
value stored in some of the corpse's unused header fields (format &
identityHash).  0 marks the end of the list.  Of course this doesn't work if an
object is the first object in newSpace as this offset will be computed to be
zero, improperly teminating the list!

Fix this by adding one to the corpse offset and substracting 1 when decoding.

Implement a simple policy to deal with the fact that typically heap growth
happens during tenuring, not after a failed allocation.  If, after scavenging,
the heap has grown by some factor of its size at the previous global GC, do a
global GC.  Default the factor to 0.333333.  Provide VM parameter access to this
value:

55	ratio of growth and image size at or above which a GC will be performed post scavenge

Add a printFreeChunks debugging utility that sweeps
through memory instead of traversing the free lists.

Add a Newspeak linux Spur Cog build.