CogVM source as per VMMaker.oscog-eem.746
Implement a peephole in the Spur Cogit for an indirection
vector initialized with a single value  Avoid initializing the
slot in the array to nil and instead initialize it with the value.

Refactor setting byte1, byte2 & byte3 into loadSubsequentBytesForDescriptor:at:
for the peephole tryCollapseTempVectorInitializationOfSize:.

No longer inline CoInterpreter>>pre/postGCAction: for VM profiling.

Increase the number of trampoline table slots.