CogVM source as per VMMaker.oscog-eem.755
Spur:
Implement unforwarding in inlined machine code #==.  Hence rewrite the /horrible/
StackToRegisterMappingCogit>>genSpecialSelectorEqualsEquals.

Unforward method literals in cog:selector: to ensure that all
examinations of constants in the Cogit are of the actual objects.

Have StackToRegisterMappingCogit>>genPushReceiverBytecode
use ReceiverResultReg if it contains self.

Speed up isForwarded: test by using the fact that the
isForwarded class pun is a power of two.

Fix a slip in an assert in unlinkSolitaryFreeTreeNode:.

Comment out the checkFreeSpace check in copyToOldSpace:bytes:format:.

Make the assert check in Spur32BitCoMemoryManager>>freeStart: more lenient to
avoid bogus assert fails from machine code allocations.

Fix the abort sequence for RISCs.  The assert check on the sendMissCall requires
that sendMissCall means the sendMissCall. But on RISCs we must push the link reg
before calling it.  Hence we need a sendMiss label.