CogVM source as per VMMaker.oscog-eem.766
Add convenience scripts to the build.linux32x86 buids to make all configs.

Remove some compilation warnings in cogit.c.

Correct some decls in genEnsureOopInRegNotForwarded:scratchReg:

Add a pushLinkReg arg to the core trampoline generation routines.  This is true
everywhere but in the nonLocalReturn and checkForInterrupts trampolines, which on
RISCs write the link reg, and on CISCs pop the ret pc to instructionPointer.

Remove the callJumpBar: parameter; this is always true.

Remember to save and restore the link register around the
call to ceCheckProfileTick in the primitive return machinery.

Streamline the genPrimReturnEnterCogCodeEnilopmart.
Streamline genExternalizePointersForPrimitiveCall