CogVM source as per VMMaker.oscog-eem.772
Spur:
Fix fillObj: signedness for objects straddling the mid-
point of the address space (quickly affects linux).
Similarly for routines in pigCompact, to get asserts correct.

Fix printOopsFrom:to:. for objects up to endOfMemory.

Nuke unused SpurMemMgr inst var var.

Declare lastFreeChunk and firstFreeChunk correctly.
Fix numberOfForwarders: and printForwarders: for
isForwarded:'s blindness towards freeChunks.
Comment isForwarded: to be clear on the issue.

Have the segment manager pass to sqAllocateMemorySegmentOfSize: the address of
the first large enough gap in the address space, instead of the address of the
end of the first segment.  This allows e.g. linux to use MAP_FIXED and hence
get past a 128Mb limit on mmapping.

Fix bugs in isValidFreeObject: & printFreeTreeChunk: that caused bogus assert
failures.

Cog:
Fix an abort (relocating call to invalid address) ue to an over-zealous check
in relocateCallBeforeReturnPC:by:.  Since we relocate e.g. calls to primitives
there can be no effective range check there-in.

Generate the XDisplayControlPlugin with the Cog conf.  Fix a typo in a selector.

Buildland:
Fix all uses of trap in scripts, all of which lines incorrectly ended with 0.

Add convenience scripts that just make the VMs that are uploaded.