CogVM source as per VMMaker.oscog-eem.775
Rationalize the allocation check filler between V3 ObjMem and Spur.  Make it
applicable only to plugin prims and optional, via the checkAllocFiller flag.
Add a prim failure code for this, PrimErrWritePastObject.  Make the Cogit check
and fail offending ext prims if the flag is set.  Don't fill new space with the
alloc check filler if the flag is not set.

Provide a -checkpluginwrites command line flag to turn on the check.

This is good for a -49% increase in the performance of e.g.
	[1 to: 1000000000 do: [:i| {nil}]] timeToRun
on Spur.

Refactor numStrongSlotsOf:ephemeronInactiveIf: and inline it in
scavengeReferentsOf:, reusing the object format for the isWeakling test.
Provide numStrongSlotsOfWeakling: for weakling nilling, and hence arrange
that the numStrongSlotsOf:ephemeronInactiveIf: is always non-nil.

Reorganize the StackInterpreter classes when building a VMMaker image.