CogVM source as per VMMaker.oscog-eem.794
Fix the shift for divide issues in the LargeIntegersPlugin.  Add code to
generateDivide:on:indent: to spit out checking asserts if required.

Change the SmartSyntaxPluginCodeGenerator to generate
code that ifdefs out the remapOop:in: rigmarole on Spur.

Fix inline cache for Characters in Spur.

Update mksistaarchives for new build structure on Mac (only).