CogVM source as per VMMaker.oscog-eem.907
Spur:
Implement image segments.  Use Igor's idea of splitting the computation of the
transitive closure from the writing of the image segment for simplicity, even
if the two are together into a single primitive for backwards-compatibility.

Status is that raw use of the primtiives works, but use in the context of
ImageSegmentTest fails apparently because something somewhere changes the
version stamp at teh start of a segment.  Any clues appreciated.


Rename markAccessibleObjects to markAccessibleObjectsAndFireEphemerons, rename
markObjects to markObjects: flag, where flag is true if objects are known to be
unmarked before hand.  Move expungeDuplicateAndUnmarkedClasses: and
nilUnmarkedWeaklingSlots into markObjects: from globalGarbageCollect.  This is
all to ensure that the mark phase of the storeImageSegment prim does process
weaklings fully before computing the transitive closure.

Extend the object shortening code to include word objects (for the image segment
arrays) and make them deal with new space objects as well as old space objects.
Have freeObject: attempt to coalesce to avoid fragmenting the largest free
chunk all the time it is used for scratch storage).

Nuke the abortive attempt to implement the Spur
image segment code similarly to the V3 code.

Simplify isHiddenObj: for leak checking to tolerate more punned scratch objects
in new space.  Add a flag to checkForLeaks to specifically check for leaks
around the segment prims.

Change readableFormat: so that on Spur there is as
yet no acceptable compatibility version.

All:
Rewrite the two primitives to pass back an error code on failure, and rewrite
the (New)ObjectMemory code to answer informative failure codes.

Make findClassContainingMethod:startingAt: et al robust in the face of
faulted-out classes (nil method dictionaries).

No longer parenthesize BaseHeaderSize (and soon BytesPerOop and BytesPerWord)
for upcoming 64-bit changes.

Fix parenthesization bugs in validation code in some Alien plugin primtiives.

Include the AioPlugin in the Newspeak linux VMs; recent reports indicate it's
essential to good performance in recent versions of OSProcessPlugin.