CogVM source as per VMMaker.oscog-eem.1022
Fix monumental blunder in jitted store-check code.  The saved registers mask
needs to be the *or* of the value reg and the caller-saved registers, not the
*and*!!

Bow to the inevitable and add a manual override for primitive accessor depth.
Use it in at: & at:put:.  If manual override is not used then the automatically
computed depth is too deep because of context access.

Slang:
Eliminate null expressions in and: and or: if
generateDeadCode is false.

Sista:
Added super calls in conditionnal jumps generation in order not to generate
counters for optimized methods.

Added unconditionnal single bytecode 217. This bytecode can be used in two case:
- when a method is discarded, all its bytecodes are replaced by the
  unconditional trap and its cogVMState is flushed
- in optimized methods, in a branch that wants to trigger dynamic deoptimization