CogVM source as per VMMaker.oscog-eem.1060
Clip the argument in the 64-bit versions of primitiveTimesTwoPower and avoid
the compiler warning with a cast.  Be clever with the SmallFloat version to
show how convenient the SmallFloat zeros are to work with.

primitiveDoNamedPrimitiveWithArgs needs to follow the same convention re
primitiveFunctionPointer as primitiveDoPrimitiveWithArgs et al on failure,
otherwise checkForAndFollowForwardedPrimitiveState will be fatally confused.
Add an assert to check the range of accessorDepth there-in.

Linux: Make UTF-8 the default copy/paste encoding to match the help text.