CogVM source as per VMMaker.oscog-eem.1118
Work around a compiler bug in 64-bit gcc 4.4.7-4 on RedHat that elided one of
the two assignments to a corpse's header in threading it onto a weak survivor
list.  Do so by, in 64-bits, performing a single assignment to the corpse's
header, instead of two.

Mark copyToOldSpace:bytes:format: as not to be inlined so as to avoid impacting
copyAndForward's icache density.

Simplify the macros in sqMemoryAccess.h to avoid the double cast on the
oop argument of fooAtPointer[Put].