CogVM source as per VMMaker.oscog-eem.1241
Change source file output for the Cogit so that cogit.c includes cogitARMv5.c
or cogitIA32.c as appropriate, using the same style as ThreadedFFIPlugin.

Add builds for the ARMv5 Cogit.

Changed the names of register allocation methods for more explict names, for
instance, allocateOneReg -> allocateRegForStackTop.

Cogit:
Change map generation and interpretation so that the displacements are in
codeGranularity units, and make the codeGranularity 4 on ARM.  Comment
machineCodeSize and maxSize.

Introduce UnfailingPrimitive and use it to avoid generating the interpreter
primitive call in Spur the Character>>asInteger primitive.

Revise Call/Jump terminology using CallFull/JumpFull for calls/jumps anywhere in
the address space.  Makes no difference to x86 but is operative on ARM.  ARM
call/branch ranges use immediate call/jump for all intra-zone transfers, and
5 instruction sequences only for trampoline and interpreter primitive calls.

- Change the code generation of == for Spur; only check for forwarders if
  operands differ.
- Change the code generation of == for sistaCogit
- change the counter reg to be allocated instead of being a fixed abstract reg
- fix a C code gen bug in code gen of #==
- fix a bug in #== where the arg was not unforwarded

Use new register allocation in #== with V3 in order to limit register moves.