CogVM source as per VMMaker.oscog-eem.1315
Slang:
Do a more thorough job of transforming complex to:[by:]do: loops so that the
limit expression is only evaluated once, not on each iteration of the loop.

Fix regression in read-before-written code when
rewriting to only initialize simple types.

Cogit:
Fix ARM-only regression in StackToRegisterMappingCogit's
PIC aborts when the SimpleStackBasedCogit was fixed.